+++ test/service/process-natural-main.js
@@ -1,0 +1,7 @@
+"use strict";
+
+skynetcore.error("PROCESS_NATURAL_START");
+setTimeout(() => {
+    process.exitCode = 7;
+    skynetcore.command("ABORT");
+}, 10);